backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
cohere>=4.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0